+++ bfg9000/builtins/dist.py
@@ -18,7 +18,7 @@ def extra_dist(builtins, files=None, dirs=None):
     for i in iterate(files):
         builtins['generic_file'](i)
     for i in iterate(dirs):
-        builtins['directory'](i)
+        builtins['directory'](i, include='*')
 
 
 def _dist_command(backend, format, build_inputs, buildfile, env):