+++ doc/about/changes.md
@@ -10,6 +10,11 @@ in progress
 - Support Java/Scala
 - Add support for user-defined arguments
 
+### Breaking changes
+
+- `directory()` and `header_directory()` no longer automatically include all
+  files within them (pass `include='*'` for the old behavior)
+
 ---
 
 ## v0.2.0